extract media player from radio/index.js
