+++ source/views/streaming/radio/types.js
@@ -1,0 +1,7 @@
+// @flow
+
+export type Viewport = {width: number, height: number}
+
+export type HtmlAudioError = {code: number, message: string}
+
+export type PlayState = 'paused' | 'playing' | 'checking' | 'loading'
